Refactor some bits usable in several places, remove AdaptiveTimeStepper since it behaves badly, get ready to allow splittings, update tests
