+++ irksome/__init__.py
@@ -9,5 +9,4 @@ from .getForm import getForm                  # noqa: F401
 from .stage import StageValueTimeStepper      # noqa: F401
 from .deriv import Dt                         # noqa: F401
 from .stepper import TimeStepper              # noqa: F401
-from .stepper import AdaptiveTimeStepper      # noqa: F401
 from .pc import RanaBase, RanaDU, RanaLD      # noqa: F401
