+++ tests/test_accuracy.py
@@ -6,7 +6,7 @@ from firedrake import (diff, div, dx, errornorm, exp, grad,
                        SpatialCoordinate, TestFunction, UnitIntervalMesh)
 
 from irksome import Dt, TimeStepper, GaussLegendre
-from irksome.getForm import AI, IA
+from irksome.tools import AI, IA
 from ufl.algorithms import expand_derivatives
